🚚 rename: 케밥케이스로 폴더 이름 변경
