+++ components/create/template-option/BasicTemplate.js
